fix: auto connect
